+++ molo/core/middleware.py
@@ -5,6 +5,7 @@ from django.views.defaults import permission_denied
 from django_cas_ng.middleware import CASMiddleware
 from django_cas_ng.views import login as cas_login, logout as cas_logout
 from django.contrib.auth.views import login, logout
+from django.conf import settings
 # test
 from django.contrib.messages import get_messages
 
@@ -19,6 +20,12 @@ class MoloCASMiddleware(CASMiddleware):
             return cas_login(request, *view_args, **view_kwargs)
         elif view_func == cas_logout:
             return cas_logout(request, *view_args, **view_kwargs)
+
+        if settings.CAS_ADMIN_PREFIX:
+            if not request.path.startswith(settings.CAS_ADMIN_PREFIX):
+                return None
+        elif not view_func.__module__.startswith('django.contrib.admin.'):
+            return None
 
         if request.user.is_authenticated():
             if request.user.is_staff: