+++ molo/core/tests/test_models.py
@@ -9,8 +9,8 @@ from django.contrib.auth.models import User
 from molo.core.models import ArticlePage
 from molo.core import constants
 from molo.core.tests.base import MoloTestCaseMixin
-
-from wagtail.wagtailimages.tests.utils import Image, get_test_image_file
+from wagtail.wagtailimages.models import Image
+from wagtail.wagtailimages.tests.utils import get_test_image_file
 
 
 @pytest.mark.django_db
@@ -232,3 +232,74 @@ class TestModels(TestCase, MoloTestCaseMixin):
             ArticlePage.objects.filter(tags__name='love').count(), 1)
         self.assertEquals(
             ArticlePage.objects.filter(tags__name='peace').count(), 1)
+
+    def test_meta_data_tags(self):
+        User.objects.create_superuser(
+            username='testuser', password='password', email='test@email.com')
+        self.client.login(username='testuser', password='password')
+
+        post_data = {
+            'title': 'this is a test article',
+            'slug': 'this-is-a-test-article',
+            'body-count': 1,
+            'body-0-value': 'Hello',
+            'body-0-deleted': False,
+            'body-0-order': 1,
+            'body-0-type': 'paragraph',
+            'metadata_tags': 'love, happiness',
+            'action-publish': 'Publish'
+        }
+        self.client.post(
+            reverse('wagtailadmin_pages:add',
+                    args=('core', 'articlepage', self.yourmind.id, )),
+            post_data)
+        post_data.update({
+            'title': 'this is a test article2',
+            'slug': 'this-is-a-test-article-2',
+            'metadata_tags': 'peace, happiness',
+        })
+        self.client.post(
+            reverse('wagtailadmin_pages:add',
+                    args=('core', 'articlepage', self.yourmind.id, )),
+            post_data)
+
+        self.assertEquals(
+            ArticlePage.objects.filter(
+                metadata_tags__name='happiness').count(), 2)
+        self.assertEquals(
+            ArticlePage.objects.filter(
+                metadata_tags__name='love').count(), 1)
+        self.assertEquals(
+            ArticlePage.objects.filter(
+                metadata_tags__name='peace').count(), 1)
+
+    def test_social_media(self):
+
+        User.objects.create_superuser(
+            username='testuser', password='password', email='test@email.com')
+        self.client.login(username='testuser', password='password')
+
+        self.mk_article(
+            self.yourmind, title="New article",
+            social_media_title='media title',
+            social_media_description='media description',)
+
+        self.mk_article(
+            self.yourmind, title="New article2",
+            social_media_title='media title',
+            social_media_image=self.image,)
+
+        self.assertEquals(
+            ArticlePage.objects.filter(
+                social_media_title='media title').count(), 2)
+        self.assertEquals(
+            ArticlePage.objects.filter(
+                social_media_description='media description').count(), 1)
+        self.assertEquals(
+            ArticlePage.objects.filter(
+                social_media_image=self.image).count(), 1)
+
+        response = self.client.get('/english/your-mind/new-article/')
+
+        self.assertEquals(response.status_code, 200)
+        self.assertNotContains(response, 'media title')